extracted/refactored behavior module
