+++ apiserver/pom.xml
@@ -210,6 +210,11 @@
             <artifactId>facebookmessenger-impl</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>ai.labs</groupId>
+            <artifactId>behavior-impl</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
