+++ behavior-impl/src/main/java/ai/labs/behavior/impl/BehaviorGroup.java
@@ -1,0 +1,21 @@
+package ai.labs.behavior.impl;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.LinkedList;
+import java.util.List;
+
+/**
+ * @author ginccc
+ */
+@Getter
+@Setter
+public class BehaviorGroup {
+    private String name;
+    private List<BehaviorRule> behaviorRules;
+
+    BehaviorGroup() {
+        behaviorRules = new LinkedList<>();
+    }
+}
