+++ behavior-impl/src/main/java/ai/labs/behavior/impl/IBehaviorDeserialization.java
@@ -1,0 +1,10 @@
+package ai.labs.behavior.impl;
+
+import ai.labs.serialization.DeserializationException;
+
+/**
+ * @author ginccc
+ */
+public interface IBehaviorDeserialization {
+    BehaviorSet deserialize(String json) throws DeserializationException;
+}
