--- coreserver/src/main/java/ai/labs/core/behavior/IBehaviorDeserialization.java
@@ -1,10 +1,0 @@
-package ai.labs.core.behavior;
-
-import ai.labs.serialization.DeserializationException;
-
-/**
- * @author ginccc
- */
-public interface IBehaviorDeserialization {
-    BehaviorSet deserialize(String json) throws DeserializationException;
-}
