+++ runtime-definition/src/main/java/ai/labs/runtime/IExecutablePackage.java
@@ -6,7 +6,9 @@ import ai.labs.lifecycle.ILifecycleManager;
  * @author ginccc
  */
 public interface IExecutablePackage {
-    String getContext();
+    String getName();
+
+    String getDescription();
 
     ILifecycleManager getLifecycleManager();
 }
